Move Plugin interface and utilities to core
