+++ packages/core/src/cli/project/models/plugin.ts
@@ -1,4 +1,4 @@
-import { Plugin as UserPlugin } from '../../../shared/plugin'
+import { Plugin as UserPlugin } from '../../../node/plugin'
 import { Renderer } from './renderer'
 
 export type Plugin = Omit<UserPlugin, 'renderer'> & {
