+++ packages/core/src/node/plugin/adapter.ts
