+++ packages/core/src/shared/index.ts
@@ -1,6 +1,5 @@
 export { useContext } from './context'
 export * as error from './error'
 export * as fs from './fs'
-export * as plugin from './plugin'
 export * as configuration from './configuration'
 export * as target from './target'
