+++ packages/plugin-react/src/gestalt.plugin.integration.test.ts
@@ -2,7 +2,8 @@
 import ReactPlugin from './gestalt.plugin'
 import { describe, expect, test } from 'vitest'
 import { createServer } from 'vite'
-import { fs, Plugin } from '@gestaltjs/plugins'
+import { fs } from '@gestaltjs/core/cli'
+import { Plugin } from '@gestaltjs/core/node/plugin'
 import { Plugin as RollupPlugin } from 'rollup'
 import { temporary } from '@gestaltjs/testing'
 import { rollup } from 'rollup'
