+++ packages/plugin-svelte/package.json
@@ -39,7 +39,6 @@
   },
   "devDependencies": {
     "@gestaltjs/core": "0.7.8",
-    "@gestaltjs/plugins": "0.7.8",
     "svelte": "^3.48.0",
     "vite": "2.9.9",
     "rollup": "^2.73.0",
