+++ packages/plugin-svelte/src/gestalt.plugin.ts
@@ -1,4 +1,4 @@
-import { definePlugin } from '@gestaltjs/plugins'
+import { definePlugin } from '@gestaltjs/core/node/plugin'
 import { svelte } from '@sveltejs/vite-plugin-svelte'
 import { hydrate } from './renderer/hydrate'
 import { ssr } from './renderer/ssr'
