--- packages/plugins/src/runtime/index.ts
@@ -1,4 +1,0 @@
-export { error, fs, useContext } from '@gestaltjs/core/shared'
-export { define as definePlugin } from './define'
-import { plugin } from '@gestaltjs/core/shared'
-export type Plugin = plugin.Plugin
