+++ tsconfig.base.json
@@ -17,11 +17,11 @@
     "isolatedModules": false,
     "paths": {
       "@gestaltjs/core/node/path": ["../../core/src/node/path.ts"],
+      "@gestaltjs/core/node/plugin": ["../../core/src/node/plugin.ts"],
       "@gestaltjs/core/shared": ["../../core/src/shared/index.ts"],
       "@gestaltjs/core/runtime": ["../../core/src/runtime/index.ts"],
       "@gestaltjs/core/cli": ["../../core/src/cli/index.ts"],
-      "@gestaltjs/testing": ["../../testing/src/index.ts"],
-      "@gestaltjs/plugins": ["../../plugins/src/runtime/index.ts"]
+      "@gestaltjs/testing": ["../../testing/src/index.ts"]
     }
   },
   "exclude": ["**/dist/**", "**/*/vitest.config.ts", "node_modules"]
